# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de controle financeiro pessoal com autenticação JWT",
        "title": "Imperium API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tron/projects/{id}/directives": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all CIO directives for a project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tron-directives"
                ],
                "summary": "List directives for a project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by active status",
                        "name": "active",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TronDirectiveListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "CIO creates a new strategic directive for the project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tron-directives"
                ],
                "summary": "Create a new directive",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Directive data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateTronDirectiveRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.TronDirective"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tron/projects/{id}/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns agent activity logs for a TRON project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tron-logs"
                ],
                "summary": "List agent logs for a project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by agent type",
                        "name": "agent_type",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by success status",
                        "name": "success",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TronAgentLogListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateTronDirectiveRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "priority": {
                    "$ref": "#/definitions/models.TronDirectivePriority"
                },
                "repo_id": {
                    "description": "If scope is repo",
                    "type": "string"
                },
                "scope": {
                    "$ref": "#/definitions/models.TronDirectiveScope"
                }
            }
        },
...
        "models.TronAgentLog": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "e.g., \"generate_task\", \"implement\", \"review\"",
                    "type": "string"
                },
                "agent_type": {
                    "$ref": "#/definitions/models.TronAgentType"
                },
                "created_at": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "full_prompt": {
                    "description": "For debugging",
                    "type": "string"
                },
                "full_response": {
                    "description": "For debugging",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_summary": {
                    "type": "string"
                },
                "metrics": {
                    "$ref": "#/definitions/models.TronAgentLogMetrics"
                },
                "output_summary": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "reasoning": {
                    "description": "Agent's reasoning",
                    "type": "string"
                },
                "repo_id": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "task_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.TronAgentLogListResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TronAgentLog"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "models.TronAgentLogMetrics": {
            "type": "object",
            "properties": {
                "cost_usd": {
                    "type": "number"
                },
                "duration_ms": {
                    "type": "integer"
                },
                "model": {
                    "description": "claude-sonnet-4-5-20250929, etc",
                    "type": "string"
                },
                "tokens_input": {
                    "type": "integer"
                },
                "tokens_output": {
                    "type": "integer"
                }
            }
        },
...
        "models.TronAgentType": {
            "type": "string",
            "enum": [
                "orchestrator",
                "board",
                "pm",
                "dev",
                "qa",
                "integration"
            ],
            "x-enum-varnames": [
                "AgentTypeOrchestrator",
                "AgentTypeBoard",
                "AgentTypePM",
                "AgentTypeDev",
                "AgentTypeQA",
                "AgentTypeIntegration"
            ]
        },
...
        "models.TronDirective": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "content": {
                    "description": "The directive text",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "description": "Optional expiration",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "priority": {
                    "$ref": "#/definitions/models.TronDirectivePriority"
                },
                "project_id": {
                    "type": "string"
                },
                "repo_id": {
                    "description": "If scope is repo",
                    "type": "string"
                },
                "scope": {
                    "$ref": "#/definitions/models.TronDirectiveScope"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.TronDirectiveListResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "integer"
                },
                "directives": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TronDirective"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "models.TronDirectivePriority": {
            "type": "string",
            "enum": [
                "normal",
                "high",
                "critical"
            ],
            "x-enum-varnames": [
                "DirectivePriorityNormal",
                "DirectivePriorityHigh",
                "DirectivePriorityCritical"
            ]
        },
        "models.TronDirectiveScope": {
            "type": "string",
            "enum": [
                "project",
                "repo"
            ],
            "x-enum-comments": {
                "DirectiveScopeProject": "Applies to all repos",
                "DirectiveScopeRepo": "Applies to specific repo"
            },
            "x-enum-descriptions": [
                "Applies to all repos",
                "Applies to specific repo"
            ],
            "x-enum-varnames": [
                "DirectiveScopeProject",
                "DirectiveScopeRepo"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite: Bearer {seu_token_aqui}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}